// src/api/generated/getScreeningResponse.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * Cinema Platform API
 * API Gateway for Cinema Platform microservices
 * OpenAPI spec version: 1.0
 */
import type { ScreeningTheaterInfo } from './screeningTheaterInfo';
import type { ScreeningHallInfo } from './screeningHallInfo';
import type { ScreeningMovieInfo } from './screeningMovieInfo';
import type { ScreeningSeatType } from './screeningSeatType';

export interface GetScreeningResponse {
  id: string;
  startAt: string;
  endAt: string;
  theater: ScreeningTheaterInfo;
  hall: ScreeningHallInfo;
  movie: ScreeningMovieInfo;
  seatTypes: ScreeningSeatType[];
}